[Flink] Fix SonarCloud code duplication in Job base classes
