+++ streampark-flink/streampark-flink-shims/streampark-flink-shims-base/src/main/java/org/apache/streampark/flink/core/javaapi/FlinkTableJob.java
@@ -17,17 +17,11 @@
 
 package org.apache.streampark.flink.core.javaapi;
 
-import org.apache.streampark.flink.core.FlinkSqlExecutor$;
-
 import org.apache.flink.api.common.JobExecutionResult;
 import org.apache.flink.api.java.utils.ParameterTool;
 import org.apache.flink.table.api.TableEnvironment;
 
 import java.util.function.Consumer;
-
-import scala.Function1;
-import scala.runtime.AbstractFunction1;
-import scala.runtime.BoxedUnit;
 
 /**
  * Java lifecycle base class for Flink Table API (batch) jobs.
@@ -117,18 +111,7 @@ public abstract class FlinkTableJob {
      * to that type so the public surface of this class stays Scala-free.
      */
     public void sql(String sql, Consumer<String> callback) {
-        Function1<String, BoxedUnit> scalaCallback =
-            callback == null
-                ? null
-                : new AbstractFunction1<String, BoxedUnit>() {
-
-                    @Override
-                    public BoxedUnit apply(String result) {
-                        callback.accept(result);
-                        return BoxedUnit.UNIT;
-                    }
-                };
-        FlinkSqlExecutor$.MODULE$.executeSql(sql, parameter, tableEnv, scalaCallback);
+        FlinkJobSupport.executeSql(sql, parameter, tableEnv, callback);
     }
 
     /** Direct access to the underlying Table API — use this instead of delegate methods. */
@@ -141,11 +124,6 @@ public abstract class FlinkTableJob {
     }
 
     private String getAppName() {
-        // TODO: mirror EnhancerImplicit#getAppName(required = true) once ported to Java (Phase 1.3)
-        String appName = parameter.get("app.name");
-        if (appName == null || appName.isEmpty()) {
-            throw new IllegalArgumentException("[StreamPark] \"app.name\" is required");
-        }
-        return appName;
+        return FlinkJobSupport.requireAppName(parameter);
     }
 }
